refactor(eleventy): extract example shortcode to its own file
